requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9
PyMySQL==1.1.0

# WhatsApp/SMS Integration
twilio==8.10.0

# AI/Chatbot
openai==1.3.7

# Date/Time handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# PDF generation for prescriptions
reportlab==4.0.7
weasyprint==60.2

# Image processing
Pillow==10.1.0

# HTTP requests
requests==2.31.0

# JSON handling
simplejson==3.19.2

# Validation
email-validator==2.1.0

# Scheduling/Background tasks
APScheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Security
cryptography==43.0.3
bcrypt==4.1.2

# API documentation
flask-restx==1.3.0

# CORS handling
Flask-CORS==4.0.0

# Rate limiting
Flask-Limiter==3.5.0

# Monitoring and logging
sentry-sdk[flask]==1.38.0

# Development dependencies
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Production server
gunicorn==21.2.0
gevent==23.9.1

# File upload handling
Flask-Uploads==0.2.1

# QR code generation for appointments
qrcode[pil]==7.4.2

# Excel export functionality
openpyxl==3.1.2
pandas==2.1.4

# Video consultation (WebRTC)
aiortc==1.6.0

# Calendar integration
icalendar==5.0.11

# Backup utilities
boto3==1.34.0  # For AWS S3 backups

# Health checks
py-healthcheck==1.10.1
